requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
cffi==1.17.1
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
propcache==0.3.1
py-cord==2.6.1
pycparser==2.22
pygit2==1.18.0
Pygments==2.19.1
rich==14.0.0
yarl==1.20.0
